requirements.txt
Django==4.0
djangorestframework==3.14.0
environs==9.5.0
Pillow==9.4.0
python-telegram-bot==20.1
